-cp instead of -jar for java command
clojure-maven-plugin, on which this plugin is based, builds a .jar and
populates the classpath attribute of its manifest with all the classpath
elements. This is done to avoid a "command string too long" error on
Windows, but is incompatible with the way (require :abcl-contrib) works.
A change to ABCL's mechanism for finding .asd is the underlying fix.